Everything is now connected, window with slider works properly. The assembly project has been added.

+++ ImageViewer/ImageViewer/Model/Event/FilterEvent.cs
@@ -9,6 +9,6 @@ namespace ImageViewer.Model.Event
     {
         public int PresenterID { get; set; }
         public Filter.Filters Filter { get; set; }
-        public Int16 Value { get; set; }
+        public byte Value { get; set; }
     }
 }
